requirements.txt
streamlit==1.39.0
openai>=1.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
